backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
PyMySQL==1.1.1
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
redis==5.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.32.3
